fix: resolve lint errors - extract user_input elicitation to separate file
- Extract tryUserInputElicitation to kairos_next-user-input-elicitation.ts
- Remove unused prompt variable in kairos_next-previous-step.ts
- Reduce kairos_next-pow-helpers.ts from 411 to 301 lines (under 350 limit)
